Create representation_algo
