+++ src/lib.rs
@@ -24,6 +24,7 @@ mod iterators;
 mod layout_algo;
 mod matching;
 mod random_circuit;
+mod representation_algo;
 mod shortest_path;
 mod simple_path;
 mod transitivity;
@@ -40,6 +41,7 @@ use isomorphism_algo::*;
 use layout_algo::*;
 use matching::*;
 use random_circuit::*;
+use representation_algo::*;
 use shortest_path::*;
 use simple_path::*;
 use transitivity::*;
@@ -63,8 +65,6 @@ use petgraph::visit::{
     NodeCount, NodeIndexable,
 };
 
-use ndarray::prelude::*;
-use numpy::IntoPyArray;
 use rayon::prelude::*;
 
 use crate::generators::PyInit_generators;
@@ -209,94 +209,6 @@ where
     })
 }
 
-/// Return the adjacency matrix for a PyDiGraph object
-///
-/// In the case where there are multiple edges between nodes the value in the
-/// output matrix will be the sum of the edges' weights.
-///
-/// :param PyDiGraph graph: The DiGraph used to generate the adjacency matrix
-///     from
-/// :param callable weight_fn: A callable object (function, lambda, etc) which
-///     will be passed the edge object and expected to return a ``float``. This
-///     tells retworkx/rust how to extract a numerical weight as a ``float``
-///     for edge object. Some simple examples are::
-///
-///         dag_adjacency_matrix(dag, weight_fn: lambda x: 1)
-///
-///     to return a weight of 1 for all edges. Also::
-///
-///         dag_adjacency_matrix(dag, weight_fn: lambda x: float(x))
-///
-///     to cast the edge object as a float as the weight. If this is not
-///     specified a default value (either ``default_weight`` or 1) will be used
-///     for all edges.
-/// :param float default_weight: If ``weight_fn`` is not used this can be
-///     optionally used to specify a default weight to use for all edges.
-///
-///  :return: The adjacency matrix for the input dag as a numpy array
-///  :rtype: numpy.ndarray
-#[pyfunction(default_weight = "1.0")]
-#[pyo3(text_signature = "(graph, /, weight_fn=None, default_weight=1.0)")]
-fn digraph_adjacency_matrix(
-    py: Python,
-    graph: &digraph::PyDiGraph,
-    weight_fn: Option<PyObject>,
-    default_weight: f64,
-) -> PyResult<PyObject> {
-    let n = graph.node_count();
-    let mut matrix = Array2::<f64>::zeros((n, n));
-    for (i, j, weight) in get_edge_iter_with_weights(graph) {
-        let edge_weight =
-            weight_callable(py, &weight_fn, &weight, default_weight)?;
-        matrix[[i, j]] += edge_weight;
-    }
-    Ok(matrix.into_pyarray(py).into())
-}
-
-/// Return the adjacency matrix for a PyGraph class
-///
-/// In the case where there are multiple edges between nodes the value in the
-/// output matrix will be the sum of the edges' weights.
-///
-/// :param PyGraph graph: The graph used to generate the adjacency matrix from
-/// :param weight_fn: A callable object (function, lambda, etc) which
-///     will be passed the edge object and expected to return a ``float``. This
-///     tells retworkx/rust how to extract a numerical weight as a ``float``
-///     for edge object. Some simple examples are::
-///
-///         graph_adjacency_matrix(graph, weight_fn: lambda x: 1)
-///
-///     to return a weight of 1 for all edges. Also::
-///
-///         graph_adjacency_matrix(graph, weight_fn: lambda x: float(x))
-///
-///     to cast the edge object as a float as the weight. If this is not
-///     specified a default value (either ``default_weight`` or 1) will be used
-///     for all edges.
-/// :param float default_weight: If ``weight_fn`` is not used this can be
-///     optionally used to specify a default weight to use for all edges.
-///
-/// :return: The adjacency matrix for the input dag as a numpy array
-/// :rtype: numpy.ndarray
-#[pyfunction(default_weight = "1.0")]
-#[pyo3(text_signature = "(graph, /, weight_fn=None, default_weight=1.0)")]
-fn graph_adjacency_matrix(
-    py: Python,
-    graph: &graph::PyGraph,
-    weight_fn: Option<PyObject>,
-    default_weight: f64,
-) -> PyResult<PyObject> {
-    let n = graph.node_count();
-    let mut matrix = Array2::<f64>::zeros((n, n));
-    for (i, j, weight) in get_edge_iter_with_weights(graph) {
-        let edge_weight =
-            weight_callable(py, &weight_fn, &weight, default_weight)?;
-        matrix[[i, j]] += edge_weight;
-        matrix[[j, i]] += edge_weight;
-    }
-    Ok(matrix.into_pyarray(py).into())
-}
-
 fn weight_callable(
     py: Python,
     weight_fn: &Option<PyObject>,
@@ -310,89 +222,6 @@ fn weight_callable(
         }
         None => Ok(default),
     }
-}
-
-/// Compute the complement of a graph.
-///
-/// :param PyGraph graph: The graph to be used.
-///
-/// :returns: The complement of the graph.
-/// :rtype: PyGraph
-///
-/// .. note::
-///
-///     Parallel edges and self-loops are never created,
-///     even if the :attr:`~retworkx.PyGraph.multigraph`
-///     attribute is set to ``True``
-#[pyfunction]
-#[pyo3(text_signature = "(graph, /)")]
-fn graph_complement(
-    py: Python,
-    graph: &graph::PyGraph,
-) -> PyResult<graph::PyGraph> {
-    let mut complement_graph = graph.clone(); // keep same node indexes
-    complement_graph.graph.clear_edges();
-
-    for node_a in graph.graph.node_indices() {
-        let old_neighbors: HashSet<NodeIndex> =
-            graph.graph.neighbors(node_a).collect();
-        for node_b in graph.graph.node_indices() {
-            if node_a != node_b
-                && !old_neighbors.contains(&node_b)
-                && (!complement_graph.multigraph
-                    || !complement_graph
-                        .has_edge(node_a.index(), node_b.index()))
-            {
-                // avoid creating parallel edges in multigraph
-                complement_graph.add_edge(
-                    node_a.index(),
-                    node_b.index(),
-                    py.None(),
-                )?;
-            }
-        }
-    }
-    Ok(complement_graph)
-}
-
-/// Compute the complement of a graph.
-///
-/// :param PyDiGraph graph: The graph to be used.
-///
-/// :returns: The complement of the graph.
-/// :rtype: :class:`~retworkx.PyDiGraph`
-///
-/// .. note::
-///
-///     Parallel edges and self-loops are never created,
-///     even if the :attr:`~retworkx.PyDiGraph.multigraph`
-///     attribute is set to ``True``
-#[pyfunction]
-#[pyo3(text_signature = "(graph, /)")]
-fn digraph_complement(
-    py: Python,
-    graph: &digraph::PyDiGraph,
-) -> PyResult<digraph::PyDiGraph> {
-    let mut complement_graph = graph.clone(); // keep same node indexes
-    complement_graph.graph.clear_edges();
-
-    for node_a in graph.graph.node_indices() {
-        let old_neighbors: HashSet<NodeIndex> = graph
-            .graph
-            .neighbors_directed(node_a, petgraph::Direction::Outgoing)
-            .collect();
-        for node_b in graph.graph.node_indices() {
-            if node_a != node_b && !old_neighbors.contains(&node_b) {
-                complement_graph.add_edge(
-                    node_a.index(),
-                    node_b.index(),
-                    py.None(),
-                )?;
-            }
-        }
-    }
-
-    Ok(complement_graph)
 }
 
 // The provided node is invalid.
